Move duplicated `getDisplayName` utility function into `src/util`
This function is not Preact version-specific, so it doesn't belong in
the {preact, preact10}-rst modules.
